move note type radio buttons into own file
