+++ moped-editor/src/views/projects/projectView/ProjectNotes/NoteInput.js
@@ -5,9 +5,6 @@ import {
   Container,
   Grid,
   FormControl,
-  Radio,
-  RadioGroup,
-  FormControlLabel,
   Paper,
 } from "@mui/material";
 import makeStyles from "@mui/styles/makeStyles";
@@ -29,6 +26,7 @@ import { LinkNode } from "@lexical/link";
 import { ListPlugin } from "@lexical/react/LexicalListPlugin";
 import { ListNode, ListItemNode } from "@lexical/list";
 import EditorTheme from "../EditorTheme";
+import NoteTypeRadioButtons from "./NoteTypeRadioButtons"
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -65,26 +63,6 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-/**
- * Defines the NoteTypeRadioButtons that allow a user to select the note type of a
- * note they are creating or editing.
- * @param {integer} defaultValue - The note type id that determines which button option
- * should be selected/highlighted by default
- * @param {function} onChange - Callback function that runs when the note type is changed via button click
- * @return {JSX.Element}
- * @constructor
- */
-const NoteTypeRadioButtons = ({ defaultValue, onChange }) => (
-  <RadioGroup
-    row
-    value={defaultValue}
-    onChange={onChange}
-    sx={{ color: "black" }}
-  >
-    <FormControlLabel value={1} control={<Radio />} label="Internal note" />
-    <FormControlLabel value={2} control={<Radio />} label="Status update" />
-  </RadioGroup>
-);
 
 // Catch any errors that occur during Lexical updates and log them
 // or throw them as needed.
